requirements.txt
torch
torchvision
torchaudio
opencv-python
ultralytics
timm
pyyaml
numpy
